requirements.txt
aiohttp==3.12.12
discord.py==2.5.2
beautifulsoup4==4.13.4
python-dotenv==1.1.0
